+++ app/api/routes/__init__.py
@@ -2,7 +2,17 @@
 
 from fastapi import APIRouter
 
-from . import bot, conversations, files, health_data, images, login, users, utils
+from . import (
+    bot,
+    conversations,
+    files,
+    health_data,
+    images,
+    login,
+    notifications,
+    users,
+    utils,
+)
 
 router = APIRouter()
 
@@ -14,3 +24,4 @@ router.include_router(health_data.router)
 router.include_router(users.router)
 router.include_router(images.router)
 router.include_router(files.router)
+router.include_router(notifications.router)