+++ app/api/routes/bot.py
@@ -1,17 +1,19 @@
 """Routes for manage bot conversations."""
 
+import re
 from datetime import datetime
 from logging import getLogger
 
-from fastapi import APIRouter, Depends, HTTPException, Query, Response
+from fastapi import APIRouter, Body, Depends, HTTPException, Query, Response
 
 from app import crud
 from app.api.dependencies.database import get_db
 from app.api.dependencies.security import get_current_user
-from app.core.bot import get_df_response
+from app.core.bot import get_df_response, response_to_question
+from app.core.config import settings
 from app.core.health_data import hdprocessor
 from app.models import User
-from app.schemas.bot import Msg
+from app.schemas.bot import QuestionResponse
 from app.schemas.conversation import ConversationCreate, ConversationOut
 from app.schemas.health_data import HealthDataCreate, HealthDataUpdate
 
@@ -24,21 +26,62 @@ logger = getLogger(__name__)
 @router.post(
     "/process-msg",
     response_model=ConversationOut,
-    responses={500: {"description": "HealthData was not saved before processing"}},
+    responses={
+        500: {"description": "HealthData was not saved before processing"},
+        400: {
+            "description": "The user has passed both msg and question_response or none of them"
+        },
+    },
     summary="Process user message",
 )
 def bot_message_post(
     *,
     db=Depends(get_db),
-    input_pack: Msg,
     user: User = Depends(get_current_user),
     lang: str = Query("en"),
     response: Response,
+    msg: str = Body(None),
+    question_response: QuestionResponse = Body(None),
 ):
-    """Sends a message to the bot and returns the response back."""
+    """Sends a message to the bot and returns the response back. You need to pass
+    'msg' or 'question_response', but no both.
+
+    The final intent includes two headers:
+
+    - **X-Problems-Parsed** - list of parsed problems
+    - **X-Health-Data-Result** - result of the health data algorithm
+    """
+
+    if msg and question_response:
+        raise HTTPException(400, "'msg' and 'question_response' are mutually exlusive")
+    if not msg and not question_response:
+        raise HTTPException(400, "You need to pass 'msg' or 'question_response'")
+
     health_data = None
-    message = input_pack.msg
+    message = msg
     logger.debug("User's message: %r", message)
+
+    if question_response:
+        return ConversationOut(
+            display_type="default",
+            bot_msg=response_to_question(question_response, lang),
+            intent="notification-question-response",
+            user_msg=question_response.user_response,
+            user_id=user.id,
+        )
+
+    notification_question_match = re.search(settings.bot_question_message_flag, message)
+
+    if notification_question_match:
+        logger.debug("Notification question detected, returning echo")
+        message = re.sub(settings.bot_question_message_flag, "", message)
+        return ConversationOut(
+            display_type="yes_no",
+            bot_msg=message,
+            intent="notification-question-echo",
+            user_msg=message,
+            user_id=user.id,
+        )
 
     df_resp = get_df_response(user.id, message, lang)
     logger.debug("Bot response: %r", df_resp)
@@ -64,11 +107,11 @@ def bot_message_post(
             ) from exc
 
         result = hdprocessor.process_health_data(filled_hd)
-        problem_explanation = hdprocessor.gen_report(
-            hdprocessor.classify_problems(result), lang=lang
-        )
-        df_resp.bot_msg = f"{df_resp.bot_msg}.\n{problem_explanation}"
-        response.headers["health-data-result"] = result.json()
+        problems = hdprocessor.classify_problems(result)
+        problem_explanation = hdprocessor.gen_report(problems, lang=lang)
+        df_resp.bot_msg = f"{df_resp.bot_msg}\n{problem_explanation}"
+        response.headers["X-Problems-Parsed"] = problems.json()
+        response.headers["X-Health-Data-Result"] = result.json()
 
     conversation = ConversationCreate(
         user_msg=message, user_id=user.id, **df_resp.dict()