+++ app/schemas/health_data.py
@@ -1,7 +1,23 @@
 from datetime import datetime
-from typing import Optional
+from typing import List, Optional
 
 from pydantic import BaseModel, confloat, conint
+from typing_extensions import Literal
+
+
+class ClassifiedProblem(BaseModel):
+    name: str
+    severity: Literal["light", "serious"]
+
+
+class ClassifiedProblemList(BaseModel):
+    __root__: List[ClassifiedProblem]
+
+    def __iter__(self):
+        return iter(self.__root__)
+
+    def __bool__(self):
+        return bool(self.__root__)
 
 
 class QuestionCoefficients(BaseModel):