+++ app/schemas/notification.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class QuestionNotification(BaseModel):
+    content: str
+    id: str